python/requirements.txt
pynacl==1.3.0
requests==2.23.0
ruamel.yaml==0.16.10
